requirements.txt
asgiref==3.5.0
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.0
psycopg2==2.9.3
PyJWT==1.7.1
pymongo==3.12.3
pytz==2021.3
sqlparse==0.2.4
tzdata==2021.5
